refactor(cursor): split tool-definitions into naming, schema, and guidance leaves (split S04 L1/5)
